Moved atlas_gen to separate package
