+++ .travis.yml
@@ -48,27 +48,11 @@ jobs:
       before_install: source travis/install_unix.sh
       script: bash travis/test.sh
     - stage: test
-      name: "Windows Python 3.6"
-      language: shell
-      env: TRAVIS_PYTHON_VERSION=3.6
-      os: windows
-      before_install: source travis/install_windows.sh
-      script: bash travis/test.sh
-    - stage: test
       name: "Windows Python 3.7"
       language: shell
       env: TRAVIS_PYTHON_VERSION=3.7
       os: windows
       before_install: source travis/install_windows.sh
-      script: bash travis/test.sh
-    - stage: test
-      name: "macOS Python 3.6"
-      os: osx
-      language: shell
-      env:
-      - TRAVIS_PYTHON_VERSION=3.6
-      - MINICONDA_URL=https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh
-      before_install: source travis/install_unix.sh
       script: bash travis/test.sh
     - stage: test
       name: "macOS Python 3.7"
