+++ atlas_gen/__init__.py
@@ -1,0 +1,1 @@
+from atlas_gen import mesh_utils
