+++ atlas_gen/atlas_scripts/fishatlas_atlas.py
@@ -1,4 +1,7 @@
-from fishatlas_utils import add_path_inplace, collect_all_inplace
+from atlas_gen.atlas_scripts.fishatlas_utils import (
+    add_path_inplace,
+    collect_all_inplace,
+)
 
 from pathlib import Path
 import tempfile
