+++ atlas_gen/metadata_utils.py
