+++ atlas_gen/structure_json_to_csv.py
