--- brainatlas_api/atlas_gen/__init__.py
@@ -1,10 +1,0 @@
-from brainatlas_api.atlas_gen.wrapup import wrapup_atlas_from_dir
-from brainatlas_api.atlas_gen.stacks import save_anatomy, save_annotation
-from brainatlas_api.atlas_gen.structures import (
-    get_structure_children,
-    get_structure_terminal_nodes,
-)
-
-from brainatlas_api.atlas_gen import descriptors
-from brainatlas_api.atlas_gen import volume_utils
-from brainatlas_api.atlas_gen import mesh_utils
