+++ travis/test.sh
@@ -1,4 +1,5 @@
-pip install -e .[dev]
+pip install -e .
+pip install black
 conda info -a
 black ./ -l 79 --target-version py37 --check
 pytest --cov=brainatlas_api
